quiz_creator/requirements.txt
Pillow
pdf2image
